work 24 march

+++ det_sols_from_polynomial/f0poly_sols_clean.py
@@ -5,8 +5,8 @@ parser = argparse.ArgumentParser()
 
 parser.add_argument('pi1', type=float, help='site 1 quality')
 parser.add_argument('pi2', type=float, help='site 2 quality')
-parser.add_argument('q1', type=int, help='site 1 quality')
-parser.add_argument('q2', type=int, help='site 2 quality')
+parser.add_argument('q1', type=float, help='site 1 quality')
+parser.add_argument('q2', type=float, help='site 2 quality')
 parser.add_argument('l', type=float, help='interdependence (lambda)')
 
 parser.add_argument("-v", "--verbose", help="increase output verbosity",
